removed recursion

+++ driver.cpp
@@ -6,7 +6,7 @@ int main() {
     unsigned depth = 2;
     point_t  pathStart;
     point_t  pathTarget;
-    unsigned score;
+    float    score;
 
     planner.bestPath(depth, pathStart, &pathTarget, &score);
 }
